+++ R/plot.estimate_density.R
@@ -1,11 +1,19 @@
+#' @importFrom insight clean_parameters
 #' @importFrom dplyr group_by mutate ungroup select one_of n
 #' @export
-data_plot.estimate_density <- function(x, ...) {
+data_plot.estimate_density <- function(x, model = NULL, ...) {
   dataplot <- x
 
   if (!"Parameter" %in% names(dataplot)) {
     dataplot$Parameter <- "Distribution"
   }
+
+  # add component and effects columns
+  if (!is.null(model)) {
+    dataplot <- merge(dataplot, insight::clean_parameters(model), by = "Parameter")
+  }
+
+  dataplot <- .fix_facet_names(dataplot)
 
   attr(dataplot, "info") <- list("xlab" = "Values",
                                  "ylab" = "Density",
@@ -26,7 +34,7 @@ data_plot.estimate_density <- function(x, ...) {
 #' @importFrom rlang .data
 #' @importFrom ggridges geom_ridgeline
 #' @export
-plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, grid = FALSE, priors = FALSE, priors_alpha = .4, size = 1, ...) {
+plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, n_columns = 1, priors = FALSE, priors_alpha = .4, size = .9, ...) {
   # save model for later use
   model <- tryCatch(
     {
@@ -40,8 +48,14 @@ plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, g
 
 
   if (!"data_plot" %in% class(x)) {
-    x <- data_plot(x, ...)
+    x <- data_plot(x, model = model, ...)
   }
+
+  if ((!"Effects" %in% names(x) || length(unique(x$Effects)) <= 1) &&
+      (!"Component" %in% names(x) || length(unique(x$Component)) <= 1)) n_columns <- NULL
+
+  # get labels
+  labels <- .clean_parameter_names(x$Parameter, grid = !is.null(n_columns))
 
   # remove intercept from output, if requested
   x <- .remove_intercept(x, show_intercept = show_intercept)
@@ -54,7 +68,8 @@ plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, g
         color = .data$Parameter
       )) +
       geom_line(size = size) +
-      add_plot_attributes(x)
+      add_plot_attributes(x) +
+      scale_color_flat(labels = labels)
   } else {
     p <- x %>%
       ggplot(aes(
@@ -79,11 +94,22 @@ plot.see_estimate_density <- function(x, stack = TRUE, show_intercept = FALSE, g
     p <- p + add_plot_attributes(x)
   }
 
-  if ("Effects" %in% names(x) && isTRUE(grid)) {
-    if ("Component" %in% names(x))
-      p <- p + facet_wrap(~ Effects + Component, scales = "free")
-    else
-      p <- p + facet_wrap(~ Effects, scales = "free")
+
+  if (length(unique(x$Parameter)) == 1 || isTRUE(stack)) {
+    p <- p + scale_y_continuous(breaks = NULL, labels = NULL)
+  } else {
+    p <- p + scale_y_discrete(labels = labels)
+  }
+
+
+  if (!is.null(n_columns)) {
+    if ("Component" %in% names(x) && "Effects" %in% names(x)) {
+      p <- p + facet_wrap(~ Effects + Component, scales = "free", ncol = n_columns)
+    } else if ("Effects" %in% names(x)) {
+      p <- p + facet_wrap(~ Effects, scales = "free", ncol = n_columns)
+    } else if ("Component" %in% names(x)) {
+      p <- p + facet_wrap(~ Component, scales = "free", ncol = n_columns)
+    }
   }
 
   p