+++ R/plot.hdi.R
@@ -1,17 +1,29 @@
 #' @importFrom dplyr group_by mutate ungroup select one_of n
 #' @export
-data_plot.hdi <- function(x, data = NULL, grid = TRUE, ...) {
-  .data_plot_hdi(x, data, grid)
+data_plot.hdi <- function(x, data = NULL, show_intercept = FALSE, ...) {
+  .data_plot_hdi(x = x, data = data, show_intercept = show_intercept)
 }
 
 #' @export
 data_plot.bayestestR_hdi <- data_plot.hdi
 
+#' @export
+data_plot.bayestestR_eti <- data_plot.hdi
+
+
 
 #' @keywords internal
-.data_plot_hdi <- function(x, data = NULL, grid = TRUE, parms = NULL, ...) {
+.data_plot_hdi <- function(x, data = NULL, parms = NULL, show_intercept = FALSE, ...) {
   if (is.null(data)) {
     data <- .retrieve_data(x)
+  }
+
+  if (inherits(x, "bayestestR_hdi")) {
+    legend_title <- "HDI"
+    plot_title <- "Highest Density Interval (HDI)"
+  } else {
+    legend_title <- "CI"
+    plot_title <- "Credible Interval (CI)"
   }
 
   if (!is.null(parms))
@@ -71,24 +83,34 @@ data_plot.bayestestR_hdi <- data_plot.hdi
   cn <- intersect(c("x", "y", "height", "fill", "Effects", "Component"), colnames(dataplot))
   dataplot <- dataplot[, cn, drop = FALSE]
 
-  # clean cryptic names
-  if (grid) {
-    dataplot$y <- .clean_parameter_names(dataplot$y)
-    if (!is.null(levels_order)) levels_order <- .clean_parameter_names(levels_order)
-  }
-
   if (!is.null(levels_order)) {
     dataplot$y <- factor(dataplot$y, levels = levels_order)
   }
 
+  groups <- unique(dataplot$y)
+  if (!show_intercept) {
+    dataplot <- .remove_intercept(dataplot, column = "y", show_intercept)
+    groups <- unique(setdiff(groups, .intercepts()))
+  }
+
+  if (length(groups) == 1) {
+    ylab <- groups
+    dataplot$y <- 0
+  } else {
+    ylab <- "Parameters"
+  }
+
+  dataplot <- .fix_facet_names(dataplot)
+
   attr(dataplot, "info") <- list("xlab" = "Possible parameter values",
-                                 "ylab" = "Parameters",
-                                 "legend_fill" = "HDI",
-                                 "title" = "Highest Density Interval (HDI)")
+                                 "ylab" = ylab,
+                                 "legend_fill" = legend_title,
+                                 "title" = plot_title)
 
   class(dataplot) <- c("data_plot", "see_hdi", class(dataplot))
   dataplot
 }
+
 
 
 
@@ -139,12 +161,17 @@ data_plot.bayestestR_hdi <- data_plot.hdi
 #' @importFrom rlang .data
 #' @rdname data_plot
 #' @export
-plot.see_hdi <- function(x, data = NULL, show_intercept = FALSE, grid = TRUE, ...) {
+plot.see_hdi <- function(x, data = NULL, show_intercept = FALSE, n_columns = 1, ...) {
   if (!"data_plot" %in% class(x)) {
-    x <- data_plot(x, data = data)
+    x <- data_plot(x, data = data, show_intercept = show_intercept)
   }
 
-  x <- .remove_intercept(x, column = "y", show_intercept)
+  # check if we have multiple panels
+  if ((!"Effects" %in% names(x) || length(unique(x$Effects)) <= 1) &&
+      (!"Component" %in% names(x) || length(unique(x$Component)) <= 1)) n_columns <- NULL
+
+  # get labels
+  labels <- .clean_parameter_names(x$y, grid = !is.null(n_columns))
 
   p <- x %>%
     as.data.frame() %>%
@@ -159,14 +186,26 @@ plot.see_hdi <- function(x, data = NULL, show_intercept = FALSE, grid = TRUE, ..
     geom_vline(xintercept = 0, linetype = "dotted") +
     add_plot_attributes(x)
 
+  if (length(unique(x$y)) == 1  && is.numeric(x$y)) {
+    p <- p + scale_y_continuous(breaks = NULL, labels = NULL)
+  } else {
+    p <- p + scale_y_discrete(labels = labels)
+  }
 
-  if ("Effects" %in% names(x) && isTRUE(grid)) {
-    if ("Component" %in% names(x))
-      p <- p + facet_wrap(~ Effects + Component, scales = "free")
-    else
-      p <- p + facet_wrap(~ Effects, scales = "free")
+
+  if (!is.null(n_columns)) {
+    if ("Component" %in% names(x) && "Effects" %in% names(x)) {
+      p <- p + facet_wrap(~ Effects + Component, scales = "free", ncol = n_columns)
+    } else if ("Effects" %in% names(x)) {
+      p <- p + facet_wrap(~ Effects, scales = "free", ncol = n_columns)
+    } else if ("Component" %in% names(x)) {
+      p <- p + facet_wrap(~ Component, scales = "free", ncol = n_columns)
+    }
   }
 
   p
 }
 
+
+#' @export
+plot.see_eti <- plot.see_hdi