+++ R/utils.R
@@ -9,16 +9,6 @@ magrittr::`%>%`
 
 
 
-.remove_intercept <- function(x, column = "Parameter", show_intercept) {
-  if (!show_intercept) {
-    remove <- which(x[[column]] %in% c("Intercept", "zi_Intercept", "(Intercept)", "b_Intercept", "b_zi_Intercept"))
-    if (length(remove)) x <- x[-remove, ]
-  }
-  x
-}
-
-
-
 .normalize <- function(x) {
   as.vector((x - min(x, na.rm = TRUE)) / diff(range(x, na.rm = TRUE), na.rm = TRUE))
 }
@@ -70,18 +60,82 @@ magrittr::`%>%`
 
 
 
-.clean_parameter_names <- function(params) {
+#' @importFrom stats setNames
+.clean_parameter_names <- function(params, grid = FALSE) {
+
+  params <- unique(params)
+  labels <- params
+
   # clean parameters names
   params <- gsub("(b_|bs_|bsp_|bcs_)(.*)", "\\2", params, perl = TRUE)
+  params <- gsub("^zi_(.*)", "\\1 (Zero-Inflated)", params, perl = TRUE)
+  params <- gsub("(.*)_zi$", "\\1 (Zero-Inflated)", params, perl = TRUE)
   # clean random effect parameters names
-  params <- gsub("r_(.*)\\.(.*)\\.", "\\1", params)
-  params <- gsub("b\\[\\(Intercept\\) (.*)\\]", "\\1", params)
-  params <- gsub("b\\[(.*) (.*)\\]", "\\2", params)
+  params <- gsub("r_(.*)\\.(.*)\\.", "(re) \\1", params)
+  params <- gsub("b\\[\\(Intercept\\) (.*)\\]", "(re) \\1", params)
+  params <- gsub("b\\[(.*) (.*)\\]", "(re) \\2", params)
   # clean smooth terms
-  params <- gsub("^smooth_sd\\[(.*)\\]", "\\1", params)
-  params <- gsub("^sds_", "\\1", params)
+  params <- gsub("^smooth_sd\\[(.*)\\]", "\\1 (smooth)", params)
+  params <- gsub("^sds_", "\\1 (Smooth)", params)
   # remove ".1" etc. suffix
   params <- gsub("(.*)(\\.)(\\d)$", "\\1 \\3", params)
+  # fix zero-inflation part in random effects
+  params <- gsub("(.*)__zi\\s(.*)", "\\1 \\2 (Zero-Inflated)", params, perl = TRUE)
+  # fix temporary random effects token
+  params <- gsub("\\(re\\)\\s(.*)", "\\1 (Random)", params, perl = TRUE)
 
-  params
+  if (grid) {
+    params <- trimws(gsub("(Zero-Inflated)", "", params, fixed = TRUE))
+    params <- trimws(gsub("(Random)", "", params, fixed = TRUE))
+  } else {
+    params <- gsub("(Zero-Inflated) (Random)", "(Random, Zero-Inflated)", params, fixed = TRUE)
+  }
+
+  stats::setNames(params, labels)
+}
+
+
+
+.fix_facet_names <- function(x) {
+  if ("Component" %in% names(x)) {
+    x$Component <- as.character(x$Component)
+    if (!"Effects" %in% names(x)) {
+      x$Component[x$Component == "conditional"] <- "Conditional"
+      x$Component[x$Component == "zero_inflated"] <- "Zero-Inflated"
+    } else {
+      x$Component[x$Component == "conditional"] <- "(Conditional)"
+      x$Component[x$Component == "zero_inflated"] <- "(Zero-Inflated)"
+    }
+  }
+  if ("Effects" %in% names(x)) {
+    x$Effects <- as.character(x$Effects)
+    x$Effects[x$Effects == "fixed"] <- "Fixed Effects"
+    x$Effects[x$Effects == "random"] <- "Random Effects"
+  }
+  x
+}
+
+
+
+.intercepts <- function() {
+  c("(intercept)_zi", "intercept (zero-inflated)", "intercept", "zi_intercept", "(intercept)", "b_intercept", "b_zi_intercept")
+}
+
+
+.has_intercept <- function(x) {
+  tolower(x) %in% .intercepts() | grepl("^intercept", tolower(x))
+}
+
+
+.in_intercepts <- function(x) {
+  tolower(x) %in% .intercepts() | grepl("^intercept", tolower(x))
+}
+
+
+.remove_intercept <- function(x, column = "Parameter", show_intercept) {
+  if (!show_intercept) {
+    remove <- which(.in_intercepts(x[[column]]))
+    if (length(remove)) x <- x[-remove, ]
+  }
+  x
 }